# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo App API",
    "description": "Todo App REST API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    { "name": "auth", "description": "" },
    { "name": "task-lists", "description": "" },
    { "name": "tasks", "description": "" },
    { "name": "health", "description": "" }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "example": "password123"
          }
        },
        "required": ["email", "password"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "password123"
          }
        },
        "required": ["email", "password"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "email", "createdAt", "updatedAt"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["accessToken", "user"]
      },
      "TaskList": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "order": { "type": "number" },
          "userId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "order", "userId", "createdAt", "updatedAt"]
      },
      "CreateTaskListDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255, "example": "My Tasks" }
        },
        "required": ["name"]
      },
      "UpdateTaskListDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "example": "Updated List Name"
          }
        },
        "required": ["name"]
      },
      "UpdateTaskListOrderDto": {
        "type": "object",
        "properties": {
          "order": { "type": "number", "minimum": 0, "example": 2 }
        },
        "required": ["order"]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "completed": { "type": "boolean" },
          "order": { "type": "number" },
          "taskListId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "title", "completed", "order", "taskListId", "createdAt", "updatedAt"]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "Buy groceries"
          },
          "notes": { "type": "string", "example": "Milk, eggs, bread" }
        },
        "required": ["title"]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 500 },
          "notes": { "type": "string", "nullable": true },
          "completed": { "type": "boolean" }
        }
      },
      "UpdateTaskOrderDto": {
        "type": "object",
        "properties": {
          "order": { "type": "number", "minimum": 0, "example": 2 }
        },
        "required": ["order"]
      },
      "MoveTaskDto": {
        "type": "object",
        "properties": {
          "targetListId": { "type": "string", "example": "550e8400-e29b-41d4-a716-446655440000" },
          "order": { "type": "number", "minimum": 0 }
        },
        "required": ["targetListId"]
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "summary": "Health check",
        "tags": ["health"],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthResponse" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with email and password",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AuthResponse" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/task-lists": {
      "get": {
        "operationId": "TaskListsController_findAll",
        "summary": "Get all task lists for the authenticated user",
        "tags": ["task-lists"],
        "security": [{ "bearer": [] }],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TaskList" }
                    },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "TaskListsController_create",
        "summary": "Create a new task list",
        "tags": ["task-lists"],
        "security": [{ "bearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskListDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/TaskList" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/task-lists/{id}": {
      "patch": {
        "operationId": "TaskListsController_update",
        "summary": "Update task list name",
        "tags": ["task-lists"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskListDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/TaskList" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "TaskListsController_delete",
        "summary": "Delete a task list",
        "tags": ["task-lists"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "" } }
      }
    },
    "/task-lists/{id}/order": {
      "patch": {
        "operationId": "TaskListsController_updateOrder",
        "summary": "Update task list order",
        "tags": ["task-lists"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskListOrderDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/TaskList" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/task-lists/{listId}/tasks": {
      "get": {
        "operationId": "TasksController_findAll",
        "summary": "Get all tasks for a task list",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "listId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "TasksController_create",
        "summary": "Create a new task in a list",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "listId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Task" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "patch": {
        "operationId": "TasksController_update",
        "summary": "Update a task",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Task" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "TasksController_delete",
        "summary": "Delete a task",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "" } }
      }
    },
    "/tasks/{id}/order": {
      "patch": {
        "operationId": "TasksController_updateOrder",
        "summary": "Update task order within its list",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskOrderDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Task" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}/move": {
      "patch": {
        "operationId": "TasksController_move",
        "summary": "Move a task to a different list",
        "tags": ["tasks"],
        "security": [{ "bearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MoveTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Task" },
                    "timestamp": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}